requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.1
async-timeout==3.0.1
attrs==21.2.0
bandwidth-sdk==13.0.0
betterpath==0.2.2
CacheControl==0.12.6
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==8.0.3
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.13
Django==3.1.4
django-admin-honeypot==1.1.0
django-allauth==0.44.0
django-common-helpers==0.9.2
django-crontab==0.7.1
django-debug-toolbar==3.2.2
django-filter==2.4.0
django-phonenumber-field==5.0.0
djangorestframework==3.13.1
enum34==1.1.10
idna==2.10
idna-ssl==1.1.0
importlib-metadata==3.4.0
jsonpickle==0.9.6
l==0.11.0
lxml==4.6.3
Markdown==3.3.3
msgpack==1.0.2
multidict==5.2.0
numpy==1.19.5
oauthlib==3.1.0
pandas==1.1.5
phonenumbers==8.12.20
Pillow==8.0.1
plivo==4.18.1
psutil==5.8.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
signalwire==2.0.4
six==1.15.0
sqlparse==0.4.1
telnyx==1.5.0
twilio==6.43.0
typing-extensions==3.7.4.3
urllib3==1.26.2
vcversioner==2.16.0.0
vonage==2.5.5
wrapt==1.13.2
yarl==1.7.0
zipp==3.4.0
zope.interface==5.4.0
